FilterComparisonWidget

+++ demo/src/main/kotlin/com/algolia/instantsearch/demo/filter/numeric/comparison/FilterPriceView.kt
@@ -14,13 +14,13 @@ class FilterPriceView(
     val view: View,
     val price: Attribute,
     val operator: NumericOperator
-): NumberView<Int> {
+): NumberView<Long> {
 
     init {
         view.stepperText.text = "$price ${operator.raw}"
     }
 
-    override fun setComputation(computation: Computation<Int>) {
+    override fun setComputation(computation: Computation<Long>) {
         view.arrowUp.setOnClickListener {
             computation.increment()
         }
